+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/utils/TelegramAPIUrlsKeeper.kt
@@ -1,8 +1,10 @@
 package dev.inmo.tgbotapi.utils
 
+const val telegramBotAPIDefaultUrl = "https://api.telegram.org"
+
 class TelegramAPIUrlsKeeper(
     token: String,
-    hostUrl: String = "https://api.telegram.org"
+    hostUrl: String = telegramBotAPIDefaultUrl
 ) {
     val commonAPIUrl = "$hostUrl/bot$token"
     val fileBaseUrl = "$hostUrl/file/bot$token"